pyproject.toml
[tool.poetry]
name = "ska-low-mccs-pasd"
version = "0.4.0"
description = "The Monitoring, Control and Calibration Subsystem (MCCS) of the SKA Low telescope"
license = "BSD-3-Clause"
authors = ["MCCS team"]
readme = "README.md"
documentation = "https://developer.skatelescope.org/projects/ska-low-mccs-pasd"

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[tool.poetry.dependencies]
python = "~3.10"
numpy = "1.25.2"
pytango = "9.4.2"
PyYAML = "^6.0"
ska-tango-base = "^0.19.1"
ska-control-model = "0.3.1"
ska-low-mccs-common = "^0.9.1"
ska-ser-devices = "^0.1.1"
pymodbus = "^3.2.2"
watchdog = "3.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pylint = "^2.15.3"
pylint-junit = "^0.3.2"
pytest = "6.2.5"
pytest-bdd = "^6.0.1"
pytest-cov = "^3.0.0"
pytest-forked = "^1.4.0"
pytest-json-report = "^1.5.0"
pytest-mock = "^3.8.2"
pytest-repeat = "^0.9.1"
pytest-timeout = "^2.1.0"
coverage = "^6.4.2"
black = "^22.6.0"
darglint = "^1.8.1"
flake8 = "^5.0.1"
flake8-formatter-junit-xml = "^0.0.6"
flake8-builtins = "^1.5.3"
flake8-docstrings = "^1.6.0"
flake8-use-fstring = "^1.4"
flake8-rst-docstrings = "^0.2.7"
mypy = "^1.2.0"
pep8-naming = "^0.13.1"
types-PyYAML = "^6.0.11"
isort = "^5.10.1"
ska-tango-testing = "^0.5.3"

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.2"
docutils = "^0.17"
sphinx-argparse = "^0.3"
PyYAML = "^6.0"
ska-ser-sphinx-theme = "^0.1.1"
sphinx-autodoc-typehints = "^1.19"
ska-ser-devices = "^0.1.1"
sphinxcontrib-plantuml = "^0.25"
typing-extensions = "^4.6.1"       # undeclared dependency-dependency

[tool.poetry.scripts]
MccsPasdBus = "ska_low_mccs_pasd.pasd_bus.pasd_bus_device:main"
PasdBusSimulatorServer = "ska_low_mccs_pasd.pasd_bus.pasd_bus_simulator_server:main"
MccsConfigurator = "ska_low_mccs_pasd.configurator.mccs_configurator:main"
MccsSmartBox = "ska_low_mccs_pasd.smart_box.smart_box_device:main"
MccsFNDH = "ska_low_mccs_pasd.fndh.fndh_device:main"
MccsFieldStation = "ska_low_mccs_pasd.field_station.field_station_device:main"

[tool.pytest.ini_options]
testpaths = "tests/"
addopts = "--json-report --json-report-file=build/reports/report.json --cov-report html:build/htmlcov --cov-report xml:build/reports/code-coverage.xml --cov-report term:skip-covered --cov=ska_low_mccs_pasd --junitxml=build/reports/unit-tests.xml --verbose"
console_output_style = "progress"
junit_family = "legacy"
markers = [
    "SP-1931",
    "XTP-21594",
    "XTP-20301",
    "XTP-21514",
    "XTP-21515",
    "XTP-21516",
    "XTP-21517",
]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
data_file = ".coverage/.coverage"
branch = true
source = ["ska_low_mccs-pasd"]

[tool.coverage.report]
show_missing = false

[build-system]
requires = ["poetry-core>=1.1.13"]
build-backend = "poetry.core.masonry.api"
